work in progress
Former-commit-id: 7df61046b7c636e17306cac86005f54007d08ea4
Former-commit-id: 7b1db5610a426e494dbc574ecde0f58ed37592db

+++ src/ttf.h
@@ -13,16 +13,17 @@ void ttf_fini(void);
 fontp ttf_new(const char *name, int32_t pointSize, int32_t style);
 void ttf_free(fontp);
 void ttf_text_size(fontp *, const char *text,
-                   uint32_t *w, uint32_t *h,
+                   double *w, 
+                   double *h,
                    enum_fmt *fmt,
                    double scaling,
                    double advance,
                    uint8_t fixed_width);
-void ttf_putc(fontp, int32_t c, int32_t x, int32_t y, double scaling);
-void ttf_puts(fontp, const char *text, int32_t x, int32_t y, double scaling,
+void ttf_putc(fontp, int32_t c, double x, double y, double scaling);
+void ttf_puts(fontp, const char *text, double x, double y, double scaling,
               double advance,
               uint8_t fixed_width);
-void ttf_puts_no_fmt(fontp, const char *text, int32_t x, int32_t y, double scaling,
+void ttf_puts_no_fmt(fontp, const char *text, double x, double y, double scaling,
               double advance,
               uint8_t fixed_width);
 void ttf2tga(void);
